// src/types/generated/AchatResume.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type AchatResume = { id: number, date: string, fournisseur_id: number | null, reference: string | null, montant_total: number, statut: string, fournisseur_nom: string | null, statut_livraison: string, statut_paiement: string, };